Move tests to SonarAnalysisContextTest.Register.cs
